Simple C++ client
This is a first simple version of a C++ client. Right now, it handles network
connection, initial handshake and game state.  init.lua has been changed to use
the new client object for these functions. The Lua interface of the Torch
library itself has not been changed.

For now, the pure C++ layer is very thin and most of the functionality is
provided by the Lua wrapper, mostly because data from the server is sent as Lua
code.